pass configuration in from processManager -> remoteTerminalService
